+++ modules/settings/conf/conf.py
@@ -77,12 +77,16 @@ def initusersettings():
         if os.path.isfile(bad_settings_file):
             os.rename(bad_settings_file, user_settings_file)
             log.info(
-                f"Bad settings filename '{bad_settings_file}' renamed into {user_settings_file}"
+                "Bad settings filename '%s' renamed into %s",
+                bad_settings_file,
+                user_settings_file,
             )
         elif os.path.isfile(sample_settings_file):
             shutil.copy(sample_settings_file, user_settings_file)
             log.info(
-                f"No settings file found: {sample_settings_file} copied into {user_settings_file}. Set your settings."
+                "No settings file found: %s copied into %s. Set your settings.",
+                sample_settings_file,
+                user_settings_file,
             )