+++ modules/settings/conf/settings.py
@@ -10,7 +10,6 @@ Functions:
 import logging
 import pathlib
 import shutil
-import sys
 import os
 from modules.exceptions import MissingGameDirectory, UnsetGameDirectory, SettingsError
 from modules.file_utils import parseINI, readINI
@@ -70,18 +69,19 @@ def get_system_user_settings(system_settings_filename, user_settings_filename):
 
         game_dir = pathlib.Path(settings_dict["gamedir"])
         if not game_dir.is_dir():
-            raise MissingGameDirectory(f"Game directory ({game_dir}) does not exist")
+            raise MissingGameDirectory("Game directory (%s) does not exist" % game_dir)
         else:
-            logs_dir = (settings_dict["gamedir"] + "/Logs")
+            logs_dir = settings_dict["gamedir"] + "/Logs"
             subdirectories = os.listdir(logs_dir)
 
             settings_dict["zonelog"] = pathlib.PurePath(
-                game_dir, "Logs/" + subdirectories[len(subdirectories) - 1] + "/Zone.log"
+                game_dir,
+                "Logs/" + subdirectories[-1] + "/Zone.log",
             ).as_posix()
 
         log.info("Settings")
         for setting, value in settings_dict.items():
-            log.info(f" - {setting}: {value}")
+            log.info(" - %s: %s", setting, value)
     except Exception as e:
         log.error("Running without settings:", e)
 
@@ -108,7 +108,7 @@ def get_settings(settings_filename):
     try:
         settings_dict = parseINI(raw_settings["BotSettings"])
     except KeyError as kerr:
-        log.error(f"Settings file is missing section {kerr}")
+        log.error("Settings file is missing section %s", kerr)
         raise SettingsError(f"Settings file is missing section {kerr}") from kerr
 
     return settings_dict